Rename utils.ts to components.ts
